+++ .gitignore
@@ -7,3 +7,6 @@
 # cache
 **/.idea
 
+
+# pycache
+**/__pycache__/
