--- app/core/PredictLstm.py
@@ -1,4 +1,0 @@
-
-class PredictLstm:
-    def __init__(self):
-        None
