--- app/core/eco_analysis/EcoAnalysis.py
@@ -1,7 +1,0 @@
-
-class EcoAnalysis:
-    def __init__(self):
-        None
-    
-    def _load_data(self):
-        None
