--- app/core/sentimental_analysis/__init__.py
@@ -1,3 +1,0 @@
-from app.core.sentimental_analysis.bertCommunity import bertCommunity
-from app.core.sentimental_analysis.bertHeadline import bertHeadline
-from app.core.PredictLstm import lstm
