+++ app/main.py
@@ -1,11 +1,21 @@
-from fastapi import FastAPI
-from app.core.background_tasks import start_background_tasks
-from app.routers.routers import endpoints
+from fastapi import APIRouter
+from app.routers import predict
+from fastapi.templating import Jinja2Templates
+from fastapi import Request,FastAPI
+from fastapi.responses import HTMLResponse
+from dotenv import load_dotenv
 
+
+load_dotenv("../.env")
 app = FastAPI()
 
-# 백그라운드 작업 시작
-start_background_tasks()
-
 # 라우터 추가
-app.include_router(endpoints.router)
+app.include_router(predict.router)
+
+# 템플릿 설정
+templates = Jinja2Templates(directory="app/templates")
+
+# 루트 경로("/")에 대한 라우터 정의
+@app.get("/", response_class=HTMLResponse)
+async def read_root(request: Request):
+    return templates.TemplateResponse("index.html", {"request": request})
