+++ app/templates/index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Welcome to FastAPI</title>
+</head>
+<body>
+    <h1>Hello, FastAPI!</h1>
+    <p>Welcome to your first FastAPI application.</p>
+</body>
+</html>
