+++ core/PredictLstm.py
@@ -1,0 +1,4 @@
+
+class PredictLstm:
+    def __init__(self):
+        None
